venus: overlay: adjust legacy status_bar_height for old apps
Some apps still use this dimen to size their layout. That is totally
fine in theory because the framework links this dimen to
status_bar_height_portrait by default. The problem is that often times
those apps do not take status_bar_padding_top into account, so as a
result, the view is too high and cuts into the actual statusbar area.

Deliberately give this dimen a value that is approximately
status_bar_height_portrait + status_bar_padding_top in dp so that
those apps can look normal again.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>